requirements.txt
bs4
requests
Pillow
pandas